Revert "Project Panel (#1382)"
This reverts commit 5753b6b1a4c7f54028fb91e63ab63efc1dc70248.
