--- assets/test/targets/Plugins/PluginTarget/main.swift
@@ -1,9 +1,0 @@
-import PackagePlugin
-import Foundation
-
-@main
-struct MyCommandPlugin: CommandPlugin {
-    func performCommand(context: PluginContext, arguments: [String]) throws {
-        print("Plugin Target Hello World")
-    }
-}
