--- assets/test/targets/Snippets/AnotherSnippet.swift
@@ -1,1 +1,0 @@
-print("Another Snippet Hello World")
