--- assets/test/targets/Sources/ExecutableTarget/main.swift
@@ -1,1 +1,0 @@
-print("Executable Target Hello World!")
