--- assets/test/targets/Sources/LibraryTarget/Targets.swift
@@ -1,9 +1,0 @@
-// The Swift Programming Language
-// https://docs.swift.org/swift-book
-
-public func foo() {
-    print("foo")
-}
-public func bar() {
-    print("bar")
-}
