--- assets/test/targets/Tests/AnotherTests/AnotherTests.swift
@@ -1,8 +1,0 @@
-import LibraryTarget
-import XCTest
-
-class AnotherTests: XCTestCase {
-    func testExample() {
-        bar()
-    }
-}
