--- assets/test/targets/Tests/TargetsTests/TargetsTests.swift
@@ -1,8 +1,0 @@
-import LibraryTarget
-import XCTest
-
-class TargetsTests: XCTestCase {
-    func testExample() {
-        foo()
-    }
-}
