+++ src/SwiftPackage.ts
@@ -41,7 +41,7 @@ export interface Target {
     c99name: string;
     path: string;
     sources: string[];
-    type: "executable" | "test" | "library" | "snippet" | "plugin";
+    type: "executable" | "test" | "library" | "snippet";
 }
 
 /** Swift Package Manager dependency */
