+++ src/TestExplorer/TestRunner.ts
@@ -521,27 +521,12 @@ export class TestRunner {
     }
 
     /**
-     * Extracts a list of unique test Targets from the list of test items.
-     */
-    private testTargets(items: vscode.TestItem[]): string[] {
-        const targets = new Set<string>();
-        for (const item of items) {
-            const target = item.id.split(".")[0];
-            targets.add(target);
-        }
-        return Array.from(targets);
-    }
-
-    /**
      * Test run handler. Run a series of tests and extracts the results from the output
      * @param shouldDebug Should we run the debugger
      * @param token Cancellation token
      * @returns When complete
      */
     async runHandler() {
-        const testTargets = this.testTargets(this.testArgs.testItems);
-        this.workspaceContext.testsStarted(this.folderContext, this.testKind, testTargets);
-
         const runState = new TestRunnerTestRunState(this.testRun);
 
         const cancellationDisposable = this.testRun.token.onCancellationRequested(() => {
@@ -566,8 +551,6 @@ export class TestRunner {
 
         cancellationDisposable.dispose();
         await this.testRun.end();
-
-        this.workspaceContext.testsFinished(this.folderContext, this.testKind, testTargets);
     }
 
     /** Run test session without attaching to a debugger */
@@ -995,6 +978,11 @@ export class TestRunner {
                                                 runState
                                             );
                                         }
+
+                                        // show test results pane
+                                        vscode.commands.executeCommand(
+                                            "testing.showMostRecentOutput"
+                                        );
 
                                         const terminateSession =
                                             vscode.debug.onDidTerminateDebugSession(() => {
