+++ test/integration-tests/commands/dependency.test.ts
@@ -14,7 +14,10 @@
 
 import { expect } from "chai";
 import * as vscode from "vscode";
-import { PackageNode, ProjectPanelProvider } from "../../../src/ui/ProjectPanelProvider";
+import {
+    PackageDependenciesProvider,
+    PackageNode,
+} from "../../../src/ui/PackageDependencyProvider";
 import { testAssetUri } from "../../fixtures";
 import { FolderContext } from "../../../src/FolderContext";
 import { WorkspaceContext } from "../../../src/WorkspaceContext";
@@ -25,8 +28,8 @@ import { getBuildAllTask, SwiftTask } from "../../../src/tasks/SwiftTaskProvider
 
 suite("Dependency Commmands Test Suite", function () {
     // full workflow's interaction with spm is longer than the default timeout
-    // 120 seconds for each test should be more than enough
-    this.timeout(120 * 1000);
+    // 60 seconds for each test should be more than enough
+    this.timeout(60 * 1000);
 
     let defaultContext: FolderContext;
     let depsContext: FolderContext;
@@ -54,12 +57,12 @@ suite("Dependency Commmands Test Suite", function () {
     });
 
     suite("Swift: Use Local Dependency", function () {
-        let treeProvider: ProjectPanelProvider;
+        let treeProvider: PackageDependenciesProvider;
 
         setup(async () => {
             await workspaceContext.focusFolder(depsContext);
             await executeTaskAndWaitForResult((await getBuildAllTask(depsContext)) as SwiftTask);
-            treeProvider = new ProjectPanelProvider(workspaceContext);
+            treeProvider = new PackageDependenciesProvider(workspaceContext);
         });
 
         teardown(() => {
@@ -67,11 +70,8 @@ suite("Dependency Commmands Test Suite", function () {
         });
 
         async function getDependency() {
-            const headers = await treeProvider.getChildren();
-            const header = headers.find(n => n.name === "Dependencies") as PackageNode;
-            expect(header).to.not.be.undefined;
-            const children = await header.getChildren();
-            return children.find(n => n.name === "swift-markdown") as PackageNode;
+            const items = await treeProvider.getChildren();
+            return items.find(n => n.name === "swift-markdown") as PackageNode;
         }
 
         // Wait for the dependency to switch to the expected state.
